# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Soccer",
        "description": "API Soccer.",
        "contact": {
            "name": "PT XYZ",
            "url": "https://xyz.test"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get All teams",
                "operationId": "fc1c5dd9ef530faf6e8bba930cad296c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Teams retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bekasi FC"
                                                    },
                                                    "logo": {
                                                        "type": "string",
                                                        "example": "bekasi.png"
                                                    },
                                                    "founded_year": {
                                                        "type": "integer",
                                                        "example": "2004"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Bekasi"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Tambun"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Team not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Store a new team",
                "operationId": "b05c08b4a03b1d708fc94bf7dd5f8e46",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "founded_year",
                                    "address",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Bekasi FC"
                                    },
                                    "founded_year": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "2000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jl hayam wuruk no 67"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jakarta"
                                    },
                                    "logo": {
                                        "description": "Upload Team's Logo",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Team has been added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "one of fields must br required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}